feat: Create setupEventListenr and init funtions
There are couple of event listeners in controller module, so I created
setupEventListener and put all events inside of the function. I also created
init function which controller module returns to the public.
Whenever the application starts, the init function is invoked where I can
put all codes that I want to be executed as soon as the application starts.