feat: Set up delete event listener using event delegation
When an event happens, the target element is stored in the
event object. It means that I can attach an event handler to a
parent element and wait for the event to bubble up.

So, I added an event handler to the parent element of a delete
button which I am actually interested in using, but not exists
when the application begins.